+++ packages/perseus/src/util/parse-perseus-json/perseus-parsers/perseus-image-background.ts
@@ -1,9 +1,11 @@
-import {nullable} from "../general-purpose-parsers/nullable";
-import {number} from "../general-purpose-parsers/number";
-import {object} from "../general-purpose-parsers/object";
-import {optional} from "../general-purpose-parsers/optional";
-import {string} from "../general-purpose-parsers/string";
-import {union} from "../general-purpose-parsers/union";
+import {
+    nullable,
+    number,
+    object,
+    optional,
+    string,
+    union,
+} from "../general-purpose-parsers";
 
 import type {Parser} from "../parser-types";
 import type {PerseusImageBackground} from "@khanacademy/perseus";
