+++ packages/perseus/src/util/parse-perseus-json/perseus-parsers/perseus-renderer.ts
@@ -1,9 +1,11 @@
-import {array} from "../general-purpose-parsers/array";
-import {number} from "../general-purpose-parsers/number";
-import {object} from "../general-purpose-parsers/object";
-import {optional} from "../general-purpose-parsers/optional";
-import {record} from "../general-purpose-parsers/record";
-import {string} from "../general-purpose-parsers/string";
+import {
+    array,
+    number,
+    object,
+    optional,
+    record,
+    string,
+} from "../general-purpose-parsers";
 
 import {parseWidgetsMap} from "./widgets-map";
