+++ packages/perseus/src/util/parse-perseus-json/perseus-parsers/widget.ts
@@ -1,8 +1,10 @@
-import {boolean} from "../general-purpose-parsers/boolean";
-import {number} from "../general-purpose-parsers/number";
-import {object} from "../general-purpose-parsers/object";
-import {optional} from "../general-purpose-parsers/optional";
-import {string} from "../general-purpose-parsers/string";
+import {
+    boolean,
+    number,
+    object,
+    optional,
+    string,
+} from "../general-purpose-parsers";
 
 import type {WidgetOptions} from "../../../perseus-types";
 import type {Parser} from "../parser-types";
